Move Generate return logs service
